conected controlers

+++ src/controllers/car.controller.js
@@ -3,16 +3,18 @@ import { carService } from '../services/car.service.js';
 class CarController {
   async getAll(req, res, next) {
     try {
-      const result = await carService.getAll();
+      const { userId } = req.body;
+      const result = await carService.getAll({ userId });
       res.status(200).json(result);
     } catch (error) {
       next(error);
     }
   }
-  
+
   async getById(req, res, next) {
     try {
-      const result = await carService.getById();
+      const { id, userId } = req.body;
+      const result = await carService.getById({ id, userId });
       res.status(200).json(result);
     } catch (error) {
       next(error);
@@ -21,7 +23,8 @@ class CarController {
 
   async create(req, res, next) {
     try {
-      const result = await carService.create();
+      const { brand, model, registerPlate, VIN, year } = req.body;
+      const result = await carService.create({ year, brand, model, registerPlate, VIN });
       res.status(200).json(result);
     } catch (error) {
       next(error);
@@ -30,7 +33,8 @@ class CarController {
 
   async delete(req, res, next) {
     try {
-      const result = await carService.delete();
+      const { id, userId } = req.body;
+      const result = await carService.delete({ id, userId });
       res.status(200).json(result);
     } catch (error) {
       next(error);
