+++ src/controllers/order.controller.js
@@ -1,10 +1,10 @@
 import { orderService } from '../services/order.service.js';
 
 class OrderController {
-
   async getById(req, res, next) {
     try {
-      const result = await orderService.getById();
+      const { carId, userId } = req.body;
+      const result = await orderService.getById({ userId, carId });
       res.status(200).json(result);
     } catch (error) {
       next(error);
@@ -13,7 +13,8 @@ class OrderController {
 
   async create(req, res, next) {
     try {
-      const result = await orderService.create();
+      const { carId, userId, createData, workList } = req.body;
+      const result = await orderService.create({ carId, userId, createData, workList });
       res.status(200).json(result);
     } catch (error) {
       next(error);
@@ -22,7 +23,8 @@ class OrderController {
 
   async delete(req, res, next) {
     try {
-      const result = await orderService.delete();
+      const { carId, userId } = req.body;
+      const result = await orderService.delete({ userId, carId });
       res.status(200).json(result);
     } catch (error) {
       next(error);
